+++ tests/docker/install/_helper_run_install_tests.sh
@@ -63,7 +63,7 @@ echo "Installing test dependencies for Python $python_version on $pretty_name"
             TZ=America
             ln -snf /usr/share/zoneinfo/$TZ /etc/localtime && echo $TZ > /etc/timezone
             if [ "$python_version" = "3.10" ] || [ "$python_version" = "3.11" ]; then
-                # py3.10,3.11 not currently on main apt repo so use deadsnakes
+                # py3.10,3.11 not currently on main apt repo for ubuntu 18/20 so use deadsnakes
                 apt-get install -y software-properties-common
                 add-apt-repository ppa:deadsnakes/ppa -y
                 apt-get install -y \
@@ -89,17 +89,33 @@ echo "Installing test dependencies for Python $python_version on $pretty_name"
                 update-alternatives --install /usr/bin/python python "/usr/bin/python$python_version" 1
             fi
             
-            # Make sure we don't install py3.6's pip
-            # Official get-pip documentation:
-            # https://pip.pypa.io/en/stable/installation/#get-pip-py
-            wget https://bootstrap.pypa.io/get-pip.py
-            python get-pip.py
+        elif [ "$ubuntu_version" = "22.04" ]; then
+            apt-get update -y
+            TZ=America
+            ln -snf /usr/share/zoneinfo/$TZ /etc/localtime && echo $TZ > /etc/timezone
 
-            pip install --upgrade pip >/dev/null
+            apt-get install -y \
+                "python$python_version" \
+                "python$python_version-distutils" \
+                "python$python_version-dev" \
+                python3-setuptools \
+                build-essential \
+                unzip \
+                wget \
+                curl
+            update-alternatives --install /usr/bin/python python "/usr/bin/python$python_version" 1
+
         else
             echo "ERROR: Testing on Ubuntu <18.04 not supported."
             exit 1
         fi
+
+        # Make sure we don't install py3.6's pip on ubuntu
+        # Official get-pip documentation:
+        # https://pip.pypa.io/en/stable/installation/#get-pip-py
+        wget https://bootstrap.pypa.io/get-pip.py
+        python get-pip.py
+        pip install --upgrade pip >/dev/null
 
     elif grep "Amazon Linux" /etc/os-release; then
         yum update -y